requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
yt-dlp==2025.01.1
google-api-python-client==2.147.0
python-dotenv==1.0.1
